fix: restore url updates for filters
